Move component_spec and entry
