+++ crates/cgp-component-macro-lib/src/entrypoints/cgp_component.rs
@@ -2,8 +2,8 @@ use proc_macro2::TokenStream;
 use quote::ToTokens;
 use syn::ItemTrait;
 
-use crate::derive_component::component_spec::ComponentSpec;
 use crate::derive_component::derive::derive_component_with_ast;
+use crate::parse::ComponentSpec;
 
 pub fn cgp_component(attr: TokenStream, item: TokenStream) -> syn::Result<TokenStream> {
     let spec: ComponentSpec = syn::parse2(attr)?;
