+++ crates/cgp-component-macro-lib/src/getter_component/use_fields.rs
@@ -6,9 +6,9 @@ use syn::punctuated::Punctuated;
 use syn::token::Plus;
 use syn::{parse2, parse_quote, ItemImpl, ItemTrait, TypeParamBound};
 
-use crate::derive_component::component_spec::ComponentSpec;
 use crate::getter_component::getter_field::GetterField;
 use crate::getter_component::symbol::symbol_from_string;
+use crate::parse::ComponentSpec;
 
 pub fn derive_use_fields_impl(
     spec: &ComponentSpec,
