+++ crates/cgp-component-macro-lib/src/parse/entry.rs
